add data IO to separate intern/db.py file
